refactor: extract and unify station image rendering into StationImage composable
