+++ feature/stationlist/src/main/java/io/github/agimaulana/radio/feature/stationlist/component/PinnedStationRow.kt
@@ -1,8 +1,6 @@
 package io.github.agimaulana.radio.feature.stationlist.component
 
 import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.combinedClickable
 import androidx.compose.foundation.layout.Arrangement
@@ -10,7 +8,6 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
@@ -24,7 +21,6 @@ import androidx.compose.material3.contentColorFor
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
@@ -34,7 +30,6 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import coil.compose.rememberAsyncImagePainter
 import io.github.agimaulana.radio.core.design.theme.PreviewTheme
 import io.github.agimaulana.radio.feature.stationlist.R
 import io.github.agimaulana.radio.feature.stationlist.StationListViewModel.UiState.Station
@@ -106,23 +101,20 @@ private fun PinnedStationChip(
         Box(
             modifier = Modifier
                 .size(72.dp)
-                .clip(RoundedCornerShape(20.dp))
-                .background(Color.White.copy(alpha = 0.1f))
                 .then(
                     if (station.isPlaying) {
                         Modifier.border(2.dp, Color(0xFFFF4081), RoundedCornerShape(20.dp))
                     } else Modifier
                 )
         ) {
-            Image(
-                painter = rememberAsyncImagePainter(
-                    model = station.imageUrl,
-                    placeholder = painterResource(id = R.drawable.station_default),
-                    error = painterResource(id = R.drawable.station_default)
-                ),
-                contentDescription = station.name,
-                contentScale = ContentScale.Crop,
-                modifier = Modifier.fillMaxSize()
+            StationImage(
+                stationName = station.name,
+                imageUrl = station.imageUrl,
+                modifier = Modifier.align(Alignment.Center),
+                size = 72.dp,
+                cornerRadius = 20.dp,
+                placeholderSize = 36.dp,
+                imageContentScale = ContentScale.Crop,
             )
 
             // Star badge overlay
